requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
Nuitka==2.8.9
ordered-set==4.1.0
pycryptodome==3.23.0
PyQt6==6.10.1
PyQt6-Qt6==6.10.1
PyQt6_sip==13.10.3
requests==2.32.5
urllib3==2.6.2
zstandard==0.25.0
